% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMe.R
\name{sampleMe}
\alias{sampleMe}
\title{sample!}
\usage{
sampleMe(x, y)
}
\arguments{
\item{x}{A number:number}

\item{y}{A number}
}
\value{
a sample of y from x
}
\description{
sample!
}
\examples{
sample(1:10, 5)
sample(2:20, 4)
}

